requirements.txt
nextcord
google-genai
google-auth
google-api-python-client
python-dotenv
aiohttp
aiofiles
duckduckgo-search
beautifulsoup4
RestrictedPython
aiosqlite
onami
ddgs
pytz
requests
PyMuPDF
python-docx
math2docx
